+++ src/main/java/net/oneandone/pommes/repository/GithubRepository.java
@@ -44,7 +44,7 @@ public class GithubRepository extends Repository {
 
     public static GithubRepository createOpt(Environment environment, String repository, String url, PrintWriter log) throws IOException {
         if (url.startsWith(PROTOCOL)) {
-            return new GithubRepository(environment, repository, url.substring(PROTOCOL.length()), environment.lib.tokenOpt(repository, PROTOCOL));
+            return new GithubRepository(environment, repository, url.substring(PROTOCOL.length()), environment.lib.tokenOpt(repository));
         } else {
             return null;
         }